Add Category Filtering
qmlscene and the likes will be unusable.

In mobile, entering and exiting a category
momentarily show other list view at bottom.
The problem is specific to SwipeView.
The problem goes away if the new ListView
is opened from the last page of SwipeView,
so putting category at the end of the
SwipeView will hide the glitch.
Ignoring the problem for now.

Fix build failure on Linux: symbol linux
was being redefined.

Signed-off-by: atunu <atunutemp1@gmail.com>